Allow the use of single segment server names
Prior to this, if your passive_master's name was a single segment, the
regex validation would needlessly fail to accept the name.

The validate_re function was set to require a '.' character in the name
of the passive_master. This commit loosens the requirements by making
the '.' optional.

With the fix, the following types of hostnames for passive_masters are
allowed:

  * foo.company.net
  * master.puppet.company.net
  * master-a
  * master-b