Don't hard code the path to the SSH tools
The path to ssh-keygen and ssh-keyscan vary based on operatingsystem.
Prior to this commit, the paths were hard coded. While testing on CentOS
6.6, the execs would fail to run because ssh-keygen and ssh-keyscan are
not in /bin.

This commit fixes that by using the $::path fact for the execs in the
acitve::ssh class so that the system's path is searched.