Updated README and Rsync Excludes for CA Serial
Updating directions a bit to accomadate the CA serial issue.  Basically,
you sync the contents of mastera's SSL directory to masterb.  Install
PE on masterb.  THEN... you update the serial on mastera by a large
amount to make sure that none of the serials that got used in the
installation of masterb are re-used by mastera.